refactor: move internalcards para dentro de backend/diabetes
