+++ src/features/acf/backend/diabetes/externalCards/adapter.ts
@@ -1,9 +1,9 @@
+import type { AcfDashboardType } from "../../../diabetes/common/model";
 import type {
     AcfExternalCardsDescription,
     ExternalCardDataItem,
     ExternalCardDBDataItem,
-} from "../../../diabetes/backend/model";
-import type { AcfDashboardType } from "../../../diabetes/common/model";
+} from "../model";
 
 export const externalCardsDbToModel = (
     data: Array<ExternalCardDBDataItem>
