+++ src/features/acf/backend/diabetes/externalCards/controller.ts
@@ -1,5 +1,5 @@
-import type { ExternalCardDataItem } from "../../../diabetes/backend/model";
 import type { AcfDashboardType } from "../../../diabetes/common/model";
+import type { ExternalCardDataItem } from "../model";
 import { externalCardsDbToModel } from "./adapter";
 import { externalCardsDataForTeam } from "./repository";
