+++ src/features/acf/backend/diabetes/externalCards/repository.ts
@@ -1,4 +1,4 @@
-import type { ExternalCardDBDataItem } from "../../../diabetes/backend/model";
+import type { ExternalCardDBDataItem } from "../model";
 
 // TODO usar prisma para buscar os dados quando fizermos uma lista que possui cards externos
 const data = [
