+++ src/features/acf/backend/diabetes/internalCards/adapter.ts
@@ -1,4 +1,4 @@
-import type { InternalCardDataItem, InternalCardDBDataItem } from "../model";
+import type { InternalCardDataItem, InternalCardDBDataItem } from "./model";
 
 export const dbToModel = (
     data: ReadonlyArray<InternalCardDBDataItem>
