+++ src/features/acf/diabetes/backend/filters/adapter.ts
@@ -3,7 +3,10 @@ import type {
     PatientAgeRange,
     PatientStatus,
 } from "@/features/acf/diabetes/common/model";
-import type { FilterOptionsDb, FiltersOptions } from "../model";
+import type {
+    FilterOptionsDb,
+    FiltersOptions,
+} from "../../../backend/diabetes/model";
 
 // TODO mudar esses tipos para não serem hard coded
 export type DiabetesFilterOptions =
