+++ src/features/acf/diabetes/backend/table/adapter.ts
@@ -1,5 +1,10 @@
 import { isDate, parseDate } from "@/features/common/shared/time";
 import type { impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos } from "@prisma/client";
+import {
+    sortableFieldsToDb,
+    type DiabetesDbFilterItem,
+    type SortableDbField,
+} from "../../../backend/diabetes/model";
 import type {
     ConditionIdentifiedBy,
     DiabetesAcfItem,
@@ -7,11 +12,6 @@ import type {
     PatientStatus,
 } from "../../common/model";
 import type { CoeqFilters, SortableFields } from "../../common/schema";
-import {
-    sortableFieldsToDb,
-    type DiabetesDbFilterItem,
-    type SortableDbField,
-} from "../model";
 
 export const cpfOrDate = (fieldValue: string | null): Date | string | null => {
     if (fieldValue && isDate(fieldValue)) {
