+++ src/features/acf/diabetes/backend/table/repository.ts
@@ -1,7 +1,10 @@
 import type { impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos } from "@prisma/client";
 import { prisma } from "@prisma/prismaClient";
+import type {
+    DiabetesDbFilterItem,
+    SortableDbField,
+} from "../../../backend/diabetes/model";
 import type { SortOrder } from "../../common/model";
-import type { DiabetesDbFilterItem, SortableDbField } from "../model";
 const pageSize = 8;
 
 type QueryWhere = {
